Fix recursive references when using class_schema()
Fixes https://github.com/lovasoa/marshmallow_dataclass/issues/188

The @marshmallow_dataclass.dataclass decorator relies on a lazily
computed .Schema attribute/descriptor on the decorated class to solve
this issue. For class_schema() this would probably not be a valid
approach as the expectation is that its "clazz" argument would not be
modified.