A couple of fixes to make CI tests more resilient (#211)
* Explicitly set PYTHONPATH=. in mypy tests

This prevents issues when marshmallow_dataclass is installed
in editable mode. (See #209, #207.)

* Don't fail tests upon warnings from external packages.

E.g. marshmallow<3.15 generates a deprecation warning from distutils.
There's no reason that should cause a test failure for us.

* Update black target-version config.

I have no idea whether this actually affects anything, but
might as well keep it up-to-date.

* Fix typing of Union._serialize parameter

This tracks a change in type of the `attr` parameter to
`Field._serialize` made in marshmallow 3.18.0.

Here we also switch to using keyword args to typeguard.check_type
in an attempt to protect against upcoming
[changes in its signature](https://typeguard.readthedocs.io/en/latest/api.html#typeguard.check_type).

* Cherry-pick PR #207: Fix broken user types after typing-inspect 0.8.0

This is cherry-picks PR#207 by @vit-zikmund