Add support for Dict, List, set, frozenset, Tuple[T, ...], tuple[T, ...] (#221)
* Add support for Dict, List, set, frozenset

We already supported `dict` (without explicit generic type
parameters), but did not support `Dict`.  (Similarly for `list` and
`List`.)

We supported `typing.Set`, but not the builtin `set`.  (Similarly for
`typing.FrozenSet` and `frozenset`.)

We now support all of the above.

Fixes #181.

* Add support for homogeneous tuples (Tuple[T, ...])

We currently recognize `typing.Sequence[int]` as a homogeneous tuple
of ints.  This adds support for the (probably more correct) concrete
type `tuple[int, ...]` (or `typing.Tuple[int, ...]`).